+++ functions/xstate/index.js
@@ -1,0 +1,6 @@
+// xstate/index.js - Clean exports for Cloud Function
+const { billingMachine } = require('./billingMachine');
+
+module.exports = {
+  billingMachine
+};
